requirements.txt
streamlit>=1.32
python-dotenv>=1.0

# PDF ve Görsel İşleme
PyMuPDF>=1.24
Pillow>=10.0
easyocr>=1.7

# Derin Öğrenme ve VLM Modelleri
transformers<4.45.0
einops>=0.7
sentence-transformers>=3.0
huggingface-hub>=0.24

# LangChain ve RAG Mimarisi (Senin sildiğin hayati kısım)
langchain>=0.3
langchain-core>=0.3
langchain-community>=0.3
langchain-text-splitters>=0.3
langchain-chroma>=0.2
langchain-huggingface>=0.1
langchain-google-genai>=2.0
google-genai>=1.0
chromadb>=0.5

# Stabilite
numpy<2